This PR converts `annotationlib.py` into a package with  `annotationlib/__init__.py` and `annotationlib/_stringifier.py`.

Discussed here: https://discuss.python.org/t/pep-749-implementing-pep-649/54974/63

This is done in order to move the definition of `_Stringifier` into a new submodule in order to defer the import of `ast` in the main module.

The outcome of this is that `ast` should only be imported if any of the following occur:

* `get_annotations(obj, format=Format.STRING)`[^1] is used and the output is not an empty dict
* `get_annotations(obj, format=Format.FORWARDREF)` is used and there are actually forward references
* `forwardref.__forward_arg__` is called and `forwardref.__ast_node__ is not None`

**Note**: I've used a class with a `__getattr__` method as a way of deferring imports in the current PR but I'd be happy to change that to something else if there's a more standard pattern.

My machine isn't a super stable benchmarking machine so take these only as rough estimates (they're slightly different to those posted in the discuss thread as it's a different run).

This branch:
```
import time: self [us] | cumulative | imported package
...
import time:       688 |        688 |     types
import time:      3373 |       4061 |   enum
import time:       306 |        306 |   keyword
import time:       906 |       5272 | annotationlib
```

Main:
```
import time: self [us] | cumulative | imported package
...
import time:      4032 |       4032 |     _ast
import time:      1605 |       5637 |   ast
import time:       432 |        432 |     types
import time:      2027 |       2459 |   enum
import time:       126 |        126 |       itertools
import time:       205 |        205 |       keyword
import time:        77 |         77 |         _operator
import time:       341 |        418 |       operator
import time:       210 |        210 |       reprlib
import time:       357 |        357 |       _collections
import time:      1415 |       2729 |     collections
import time:       217 |        217 |     _functools
import time:      1396 |       4341 |   functools
import time:      2198 |      14634 | annotationlib
```

[^1]: or any of the similar methods such as `call_annotate_function`

<!-- gh-issue-number: gh-118761 -->
* Issue: gh-118761
<!-- /gh-issue-number -->