@AA-Turner 2025-04-02
@DavidCEllis what do you think of https://github.com/AA-Turner/cpython/commit/make_annotationlib_module?

Instead of the lazy object, it uses self-replacing functions, so that the 'lazy' cost is only paid once. I think for `ast.unparse` you could probably just get away with a local import, but I haven't benchmarked/tested this.

```py
def _Stringifier(*args, **kwds):
    # This function replaces itself with the real class when first called
    global _Stringifier
    from annotationlib._stringifier import Stringifier as _Stringifier
    return _Stringifier(*args, **kwds)
```

A