@DavidCEllis 2025-04-02
I don't mind the self replacing functions if that's a more recognisable pattern. I personally prefer the class as there's no observable placeholder object that needs to be replaced. Inspecting will only give you the actual function/module and you can't make a reference to the self-replacing function that doesn't get replaced.

The lazy object also only pays the cost once by assigning to the object after the first import - it's basically a by-hand written version of an instance of [my general lazy importer module](https://github.com/DavidCEllis/ducktools-lazyimporter).