@JelleZijlstra 2025-04-03
I'm not sure this is worth it any more:

- `import ast` is much faster with the recent changes.
- `collections` and `functools` are already imported by `typing.py`, so deferring them here doesn't feel that valuable.