@DavidCEllis 2025-04-03
Probably fair on the scale of things, although perhaps you could skip the `functools` import all together with something similar to how `dataclasses` avoids importing `typing`?

https://github.com/python/cpython/blob/d30052a4edb7d5019196757261cf573658b97f97/Lib/dataclasses.py#L808-L814

I'll note that needing `annotationlib` doesn't mean you also need `typing`. I use annotations for a dataclasses-like project where I've been very careful to keep import time down and so avoid importing `typing` at runtime. With Python 3.14, `annotationlib` on the other hand is somewhat unavoidable in the case that `FORWARDREF` is needed.