@JelleZijlstra 2025-04-04
Good idea regarding functools, I just did that myself in #132059 since the change is so simple.

I feel deferring the `ast` import in annotationlib goes too far. We need ast for both the FORWARDREF and STRING formats in `get_annotations()`, and if you're using annotationlib, you probably want one of those.

However, I do think it's realistic to defer `import annotationlib` in typing.py, I'll see if I can make that happen.