@DavidCEllis 2025-04-04
With the STRING format you do need `ast` and that's unavoidable as the fake globals always uses `_Stringifier` in that case.

With `FORWARDREF` however, `_Stringifier` will only be used if there are actual forward references that need to be replaced, so the `ast` import is only needed in this case.

Example with this PR:
```python
import sys
from annotationlib import get_annotations, Format

class NoForwardRef:
    a: int

class YesForwardRef:
    a: Any

print(get_annotations(NoForwardRef, format=Format.FORWARDREF))
print("ast" in sys.modules)

print(get_annotations(YesForwardRef, format=Format.FORWARDREF))
print("ast" in sys.modules)
```

```python
{'a': <class 'int'>}
False
{'a': ForwardRef('Any')}
True
```

I think it's likely many tools that currently get annotations directly will need to switch to using `annotationlib.get_annotations(obj, format=Format.FORWARDREF)` to avoid the *potential* `NameError` or `AttributeError` results. This doesn't necessarily mean they want *or need* `ForwardRef` in all cases, they just need to 'work' as they did with 3.13 and not raise new exceptions if someone removes `from __future__ import annotations`.

Still arguable if it goes too far with the improvement to `ast`'s own import time but I just wanted to point out that with 3.14 using `annotationlib` doesn't necessarily mean you want the elements that require `ast`.