@DavidCEllis 2025-04-11
I think this can be closed. 

The `functools` import has been deferred separately and `ast` import time has been improved so it's less of a sledgehammer to import time and `typing` is no longer being directly hit by this.

The module could still be broken up into a package, but I feel that about a fair chunk of the larger slower stdlib modules and doing what this branch does feels more 'hacky' than structured at this point.